Move expand helpers into lib/expand and split client-safe schema utils
