+++ apps/web/app/(ee)/api/partner-profile/programs/[programId]/links/route.ts
@@ -4,7 +4,7 @@ import { validatePartnerLinkUrl } from "@/lib/api/links/validate-partner-link-ur
 import { getProgramEnrollmentOrThrow } from "@/lib/api/programs/get-program-enrollment-or-throw";
 import { getExpandableRewardReferences } from "@/lib/api/rewards/link-level-rewards";
 import { parseRequestBody } from "@/lib/api/utils";
-import { parseExpandFields } from "@/lib/api/utils/get-expandable-field";
+import { parseExpandFields } from "@/lib/expand/parse-expand-fields";
 import { applyGroupUtmToLink } from "@/lib/api/utm/apply-group-utm-to-link";
 import { withPartnerProfile } from "@/lib/auth/partner";
 import { prisma } from "@/lib/prisma";
